refactor(croquis): extract import scanning into its own module
