+++ crates/vize_croquis/src/virtual_ts/generator/script.rs
@@ -10,6 +10,12 @@ use super::{
 };
 use vize_carton::{String, ToCompactString, profile};
 
+use self::import_scan::{
+    code_line_starts, find_relative_import_from_range, import_statement_end, join_statement_lines,
+};
+
+mod import_scan;
+
 impl VirtualTsGenerator {
     /// Extract setup scope info from ScopeChain.
     fn extract_setup_info(scopes: &ScopeChain) -> (Option<CompactString>, bool) {
@@ -437,485 +443,5 @@ impl VirtualTsGenerator {
     pub(crate) fn generate_emits_type(&mut self, _bindings: &BindingMetadata) {
         self.emit_line("// Emits type");
         self.emit_line("type __Emits = {};");
-    }
-}
-
-/// Report the last line index of the import statement starting at `start`.
-///
-/// Script content is processed line by line, so a wrapped import such as
-/// `import {\n  a,\n} from "./x.ts";` must be recognized as one statement.
-/// Emitting only its first line at module scope would leave the remaining
-/// lines inside the setup body, making the virtual TypeScript unparseable and
-/// erasing every type the file declares.
-///
-/// `code_starts` marks which lines begin in code context, so `import`-looking
-/// text inside a template literal or a block comment is never mistaken for a
-/// statement. Continuation lines are consumed only while the named-binding
-/// braces stay open, or when the next line opens the trailing `from` clause,
-/// so a line that merely looks like an import can never swallow the rest of
-/// the script.
-///
-/// Returns `None` when the line does not begin an import statement.
-fn import_statement_end(lines: &[&str], code_starts: &[bool], start: usize) -> Option<usize> {
-    if !code_starts.get(start).copied().unwrap_or(false) {
-        return None;
-    }
-    if !lines.get(start)?.trim().starts_with("import ") {
-        return None;
-    }
-
-    let mut statement = String::with_capacity(128);
-    for (offset, line) in lines[start..].iter().enumerate() {
-        if offset > 0 {
-            statement.push('\n');
-        }
-        statement.push_str(line.trim());
-
-        let state = scan_import_statement(&statement);
-        if state.is_complete() {
-            return Some(start + offset);
-        }
-        if state.is_open() {
-            continue;
-        }
-        // The named-binding clause is balanced but no module specifier has
-        // been seen: `import { a }\nfrom "./x";` is one statement, while
-        // anything else ends it here rather than eating unrelated code.
-        if opens_from_clause(lines.get(start + offset + 1)) {
-            continue;
-        }
-        return Some(start + offset);
-    }
-
-    Some(start)
-}
-
-/// Whether a line opens the trailing `from "…"` clause of an import.
-fn opens_from_clause(line: Option<&&str>) -> bool {
-    let Some(rest) = line.and_then(|line| line.trim_start().strip_prefix("from")) else {
-        return false;
-    };
-    rest.starts_with([' ', '\t', '"', '\''])
-}
-
-/// Join the lines of one import statement into a single logical statement.
-///
-/// Line structure is preserved: collapsing the lines into one would move any
-/// trailing `//` comment in front of the bindings that follow it and comment
-/// out the rest of the statement.
-fn join_statement_lines(lines: &[&str]) -> String {
-    if let [line] = lines {
-        return line.to_compact_string();
-    }
-
-    let mut statement = String::with_capacity(128);
-    for (offset, line) in lines.iter().enumerate() {
-        if offset > 0 {
-            statement.push('\n');
-        }
-        statement.push_str(line);
-    }
-    statement
-}
-
-/// Brace, string, and specifier state of an accumulated import statement.
-struct ImportStatementState {
-    /// Unbalanced `{` count of the named-binding clause.
-    depth: i32,
-    /// Whether a string literal is still open.
-    quote_open: bool,
-    /// Whether a closed quoted module specifier has been seen.
-    has_specifier: bool,
-}
-
-impl ImportStatementState {
-    /// A statement is complete once its braces balance, no string stays open,
-    /// and a quoted module specifier has been seen.
-    fn is_complete(&self) -> bool {
-        !self.quote_open && self.depth <= 0 && self.has_specifier
-    }
-
-    /// A statement continues onto the next line only while its named-binding
-    /// clause is still open.
-    fn is_open(&self) -> bool {
-        !self.quote_open && self.depth > 0
-    }
-}
-
-/// Scan an accumulated import statement for brace, string, and comment state.
-///
-/// Comments are skipped so that a `}` or a quote inside `// …` or `/* … */`
-/// never ends the statement early. The whole accumulated statement is rescanned
-/// on every added line, so comment state spanning lines resolves naturally.
-fn scan_import_statement(statement: &str) -> ImportStatementState {
-    let bytes = statement.as_bytes();
-    let mut depth: i32 = 0;
-    let mut quote: Option<u8> = None;
-    let mut has_specifier = false;
-    let mut line_comment = false;
-    let mut block_comment = false;
-    let mut index = 0;
-
-    while index < bytes.len() {
-        let byte = bytes[index];
-
-        if line_comment {
-            if byte == b'\n' {
-                line_comment = false;
-            }
-            index += 1;
-            continue;
-        }
-
-        if block_comment {
-            if byte == b'*' && bytes.get(index + 1) == Some(&b'/') {
-                block_comment = false;
-                index += 2;
-                continue;
-            }
-            index += 1;
-            continue;
-        }
-
-        match quote {
-            Some(open) => {
-                if byte == b'\\' {
-                    index += 2;
-                    continue;
-                }
-                if byte == open {
-                    quote = None;
-                    has_specifier = true;
-                }
-            }
-            None => match byte {
-                b'/' if bytes.get(index + 1) == Some(&b'/') => {
-                    line_comment = true;
-                    index += 2;
-                    continue;
-                }
-                b'/' if bytes.get(index + 1) == Some(&b'*') => {
-                    block_comment = true;
-                    index += 2;
-                    continue;
-                }
-                b'"' | b'\'' | b'`' => quote = Some(byte),
-                b'{' => depth += 1,
-                b'}' => depth -= 1,
-                _ => {}
-            },
-        }
-        index += 1;
-    }
-
-    ImportStatementState {
-        depth,
-        quote_open: quote.is_some(),
-        has_specifier,
-    }
-}
-
-/// Source context of a template literal being scanned.
-#[derive(Clone, Copy, PartialEq, Eq)]
-enum TemplateContext {
-    /// Inside the literal text of a template.
-    Literal,
-    /// Inside a `${ … }` expression, tracking its unbalanced `{` count.
-    Expression(u32),
-}
-
-/// Report, for every line, whether it starts in code context.
-///
-/// Import extraction is line oriented, so it would otherwise hoist an
-/// `import`-looking line out of a template literal or a block comment and
-/// corrupt both the module scope and the setup body. Strings and `//` comments
-/// cannot span lines, so only block comments and template literals carry over.
-fn code_line_starts(lines: &[&str]) -> Vec<bool> {
-    let mut starts = Vec::with_capacity(lines.len());
-    let mut block_comment = false;
-    let mut templates: Vec<TemplateContext> = Vec::new();
-
-    for line in lines {
-        starts.push(!block_comment && templates.is_empty());
-
-        let bytes = line.as_bytes();
-        let mut quote: Option<u8> = None;
-        let mut index = 0;
-
-        while index < bytes.len() {
-            let byte = bytes[index];
-
-            if block_comment {
-                if byte == b'*' && bytes.get(index + 1) == Some(&b'/') {
-                    block_comment = false;
-                    index += 2;
-                    continue;
-                }
-                index += 1;
-                continue;
-            }
-
-            if let Some(open) = quote {
-                if byte == b'\\' {
-                    index += 2;
-                    continue;
-                }
-                if byte == open {
-                    quote = None;
-                }
-                index += 1;
-                continue;
-            }
-
-            if matches!(templates.last(), Some(TemplateContext::Literal)) {
-                match byte {
-                    b'\\' => {
-                        index += 2;
-                        continue;
-                    }
-                    b'`' => {
-                        templates.pop();
-                    }
-                    b'$' if bytes.get(index + 1) == Some(&b'{') => {
-                        templates.push(TemplateContext::Expression(0));
-                        index += 2;
-                        continue;
-                    }
-                    _ => {}
-                }
-                index += 1;
-                continue;
-            }
-
-            match byte {
-                // The rest of the line is a comment.
-                b'/' if bytes.get(index + 1) == Some(&b'/') => break,
-                b'/' if bytes.get(index + 1) == Some(&b'*') => {
-                    block_comment = true;
-                    index += 2;
-                    continue;
-                }
-                b'"' | b'\'' => quote = Some(byte),
-                b'`' => templates.push(TemplateContext::Literal),
-                b'{' => {
-                    if let Some(TemplateContext::Expression(depth)) = templates.last_mut() {
-                        *depth += 1;
-                    }
-                }
-                b'}' => {
-                    if let Some(TemplateContext::Expression(depth)) = templates.last_mut() {
-                        if *depth == 0 {
-                            templates.pop();
-                        } else {
-                            *depth -= 1;
-                        }
-                    }
-                }
-                _ => {}
-            }
-            index += 1;
-        }
-    }
-
-    starts
-}
-
-/// Find a relative `from "./..."` import range with a byte scanner.
-///
-/// Script virtual-TS generation calls this for every import-looking line. A
-/// regex was measurably expensive in that loop, and the grammar we need is
-/// intentionally narrow: skip to `from`, require whitespace, accept a quoted
-/// path that begins with `.`, then return the path slice boundaries.
-fn find_relative_import_from_range(line: &str) -> Option<(usize, usize, usize, usize)> {
-    let bytes = line.as_bytes();
-    let mut search_start = 0;
-
-    while let Some(offset) = line[search_start..].find("from") {
-        let from_start = search_start + offset;
-        let mut cursor = from_start + 4;
-
-        if !bytes
-            .get(cursor)
-            .is_some_and(|byte| byte.is_ascii_whitespace())
-        {
-            search_start = cursor;
-            continue;
-        }
-
-        while bytes
-            .get(cursor)
-            .is_some_and(|byte| byte.is_ascii_whitespace())
-        {
-            cursor += 1;
-        }
-
-        let Some(quote @ (b'\'' | b'"')) = bytes.get(cursor).copied() else {
-            search_start = cursor.saturating_add(1);
-            continue;
-        };
-        let path_start = cursor + 1;
-        if bytes.get(path_start) != Some(&b'.') {
-            search_start = path_start;
-            continue;
-        }
-
-        let mut path_end = path_start + 1;
-        while bytes.get(path_end).is_some_and(|byte| *byte != quote) {
-            path_end += 1;
-        }
-        if bytes.get(path_end) != Some(&quote) {
-            return None;
-        }
-
-        return Some((from_start, path_start, path_end, path_end));
-    }
-
-    None
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        String, code_line_starts, find_relative_import_from_range, import_statement_end,
-        join_statement_lines,
-    };
-
-    fn relative_path(line: &str) -> Option<&str> {
-        find_relative_import_from_range(line).map(|(_, start, end, _)| &line[start..end])
-    }
-
-    fn statement_end(content: &str) -> Option<usize> {
-        statement_end_at(content, 0)
-    }
-
-    fn statement_end_at(content: &str, start: usize) -> Option<usize> {
-        let lines: Vec<&str> = content.lines().collect();
-        let code_starts = code_line_starts(&lines);
-        import_statement_end(&lines, &code_starts, start)
-    }
-
-    /// The module-scope text emitted for the statement starting at line 0.
-    fn joined_statement(content: &str) -> String {
-        let lines: Vec<&str> = content.lines().collect();
-        let end = statement_end(content).expect("expected an import statement");
-        join_statement_lines(&lines[0..=end])
-    }
-
-    #[test]
-    fn groups_wrapped_named_imports_into_one_statement() {
-        let content =
-            "import {\n  createScope,\n  provideScope,\n} from \"./scope.ts\";\nconst a = 1;";
-        assert_eq!(statement_end(content), Some(3));
-        assert_eq!(
-            joined_statement(content),
-            "import {\n  createScope,\n  provideScope,\n} from \"./scope.ts\";"
-        );
-    }
-
-    #[test]
-    fn groups_a_from_clause_on_its_own_line() {
-        let content = "import {\n  createScope,\n}\nfrom \"./scope.ts\";\nconst a = 1;";
-        assert_eq!(statement_end(content), Some(3));
-        assert!(joined_statement(content).ends_with("from \"./scope.ts\";"));
-    }
-
-    #[test]
-    fn keeps_line_comments_from_swallowing_later_bindings() {
-        let content =
-            "import {\n  createScope, // scope factory\n  provideScope,\n} from \"./scope.ts\";";
-        assert_eq!(statement_end(content), Some(3));
-
-        // Collapsing the lines would move `// scope factory` in front of the
-        // remaining bindings and comment out the rest of the statement.
-        let statement = joined_statement(content);
-        assert!(statement.contains("\n  provideScope,"));
-        assert!(statement.ends_with("from \"./scope.ts\";"));
-    }
-
-    #[test]
-    fn ignores_braces_and_quotes_inside_comments() {
-        let content = "import {\n  createScope, /* } it's fine */\n  provideScope,\n} from \"./scope.ts\";\nconst a = 1;";
-        assert_eq!(statement_end(content), Some(3));
-    }
-
-    #[test]
-    fn ignores_import_like_lines_inside_template_literals() {
-        let content = "const sample = `\nimport { ref } from 'vue';\n`;\nconst a = 1;";
-        assert_eq!(statement_end_at(content, 1), None);
-    }
-
-    #[test]
-    fn ignores_import_like_lines_inside_block_comments() {
-        let content = "/*\nimport { ref } from 'vue';\n*/\nconst a = 1;";
-        assert_eq!(statement_end_at(content, 1), None);
-    }
-
-    #[test]
-    fn resumes_after_a_template_literal_closes() {
-        let content =
-            "const sample = `\nimport { ref } from 'vue';\n`;\nimport { computed } from 'vue';";
-        assert_eq!(statement_end_at(content, 1), None);
-        assert_eq!(statement_end_at(content, 3), Some(3));
-    }
-
-    #[test]
-    fn tracks_nested_template_expressions() {
-        let lines: Vec<&str> = vec![
-            "const a = `${ `${ inner }` }`;",
-            "import { ref } from 'vue';",
-        ];
-        assert_eq!(code_line_starts(&lines), vec![true, true]);
-    }
-
-    #[test]
-    fn keeps_single_line_imports_untouched() {
-        let content = "import { ref } from 'vue';\nconst a = 1;";
-        assert_eq!(statement_end(content), Some(0));
-
-        let lines: Vec<&str> = content.lines().collect();
-        assert_eq!(
-            join_statement_lines(&lines[0..=0]),
-            "import { ref } from 'vue';"
-        );
-    }
-
-    #[test]
-    fn ignores_lines_that_do_not_start_an_import() {
-        assert_eq!(
-            statement_end("const a = 1;\nimport { ref } from 'vue';"),
-            None
-        );
-    }
-
-    #[test]
-    fn never_consumes_beyond_an_unclosed_import_like_line() {
-        // A stray `import`-looking line must not swallow the rest of the body.
-        let content = "import foo\nconst a = 1;\nconst b = 2;";
-        assert_eq!(statement_end(content), Some(0));
-    }
-
-    #[test]
-    fn finds_relative_import_paths() {
-        assert_eq!(
-            relative_path("import Foo from './Foo.vue';"),
-            Some("./Foo.vue")
-        );
-        assert_eq!(
-            relative_path("import type { Foo } from \"../types\";"),
-            Some("../types")
-        );
-    }
-
-    #[test]
-    fn ignores_non_relative_import_paths() {
-        assert_eq!(relative_path("import { ref } from 'vue';"), None);
-    }
-
-    #[test]
-    fn skips_from_inside_imported_names() {
-        assert_eq!(
-            relative_path("import { fromNow } from './time';"),
-            Some("./time")
-        );
     }
 }
